third_party/blink/web_tests/svg/custom/intersection-list-null-expected.txt

CONSOLE ERROR: Uncaught TypeError: Failed to execute 'checkIntersection' on 'SVGSVGElement': parameter 1 is not of type 'SVGElement'.
PASS
